+++ src/Page2Feed.Core/Model/Atom/feedEntryLink.cs
@@ -2,18 +2,20 @@
 using System.ComponentModel;
 using System.Xml.Serialization;
 
-namespace Page2Feed.Web.Model.Atom
+namespace Page2Feed.Core.Model.Atom
 {
     /// <remarks/>
     [Serializable]
     [DesignerCategory("code")]
     [XmlType(AnonymousType = true, Namespace = "http://www.w3.org/2005/Atom")]
-    public class feedLink
+    public class feedEntryLink
     {
 
         private string hrefField;
 
         private string relField;
+
+        private string typeField;
 
         /// <remarks/>
         [XmlAttribute]
@@ -29,6 +31,14 @@ namespace Page2Feed.Web.Model.Atom
         {
             get => relField;
             set => relField = value;
+        }
+
+        /// <remarks/>
+        [XmlAttribute]
+        public string type
+        {
+            get => typeField;
+            set => typeField = value;
         }
     }
 }
