+++ src/Page2Feed.Core/Model/FeedState.cs
@@ -1,4 +1,4 @@
-﻿namespace Page2Feed.Web.Model
+﻿namespace Page2Feed.Core.Model
 {
 
     public class FeedState
