+++ src/Page2Feed.Core/Model/FeedStateEx.cs
@@ -1,4 +1,4 @@
-﻿namespace Page2Feed.Web.Model
+﻿namespace Page2Feed.Core.Model
 {
 
     public class FeedStateEx : FeedState
