+++ src/Page2Feed.Web/Services/FeedService.cs
@@ -4,8 +4,8 @@ using System.Linq;
 using System.Threading.Tasks;
 using HtmlAgilityPack;
 using NLog;
-using Page2Feed.Web.Model;
-using Page2Feed.Web.Model.Atom;
+using Page2Feed.Core.Model;
+using Page2Feed.Core.Model.Atom;
 using Page2Feed.Web.Services.Interfaces;
 
 namespace Page2Feed.Web.Services
