+++ src/Page2Feed.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.28315.86
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Page2Feed.Web", "Page2Feed.Web\Page2Feed.Web.csproj", "{4DC71F0C-4740-433D-87F3-06904AED4BE3}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Page2Feed.Web", "Page2Feed.Web\Page2Feed.Web.csproj", "{4DC71F0C-4740-433D-87F3-06904AED4BE3}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Page2Feed.Core", "Page2Feed.Core\Page2Feed.Core.csproj", "{5ACE43FA-5820-4CA2-8EF8-2E76FBF5F788}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{4DC71F0C-4740-433D-87F3-06904AED4BE3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4DC71F0C-4740-433D-87F3-06904AED4BE3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4DC71F0C-4740-433D-87F3-06904AED4BE3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5ACE43FA-5820-4CA2-8EF8-2E76FBF5F788}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5ACE43FA-5820-4CA2-8EF8-2E76FBF5F788}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5ACE43FA-5820-4CA2-8EF8-2E76FBF5F788}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5ACE43FA-5820-4CA2-8EF8-2E76FBF5F788}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
